Added Error page logic with internal redirection. Changed the path_resolve and other functions to utils and changed them slightly in order to be used for error_paths.
