+++ nginx/test.conf
@@ -135,17 +135,10 @@ http
 		root /var/www/;
 		index index.html;
 		# index randomasdsdadas
+		error_page 404 /error_pages/403.html;
 		server_name first;
 		location / {
-			access_log /first_req_access.log req;
-			access_log /first_res_access.log res;
-			# access_log /first_extra_req_access.log extra_req;
-			# access_log /first_extra_res_access.log extra_res;
-			limit_except GET POST DELETE PUT {
-				deny all;
-			}
-		}
-		location /ind {
+			root /var/www/www/;
 			access_log /first_req_access.log req;
 			access_log /first_res_access.log res;
 			# access_log /first_extra_req_access.log extra_req;
@@ -154,14 +147,22 @@ http
 				deny all;
 			}
 		}
+		location /inde {
+			error_page 404 /error_pages/413.html;
+			limit_except GET {
+				deny all;
+			}
+		}
 		location /cgi/ {
+			# error_page 404 /error_pages/404.html;
+			client_max_body_size 10;
 			root /var/www/mid/;
 			# autoindex on;
 			access_log /second_req_access.log req;
 			access_log /second_res_access.log res;
 			# access_log /second_extra_req_access.log extra_req;
 			# access_log /second_extra_res_access.log extra_res;
-			limit_except GET {
+			limit_except POST GET {
 				deny all;
 			}
 			location ~ \.php$ {
