+++ public/mid/cgi/cgi_php_multi.html
@@ -8,7 +8,7 @@
         <div>
 			<h2> This is a Multipartform using PHP CGI</h2>
 			<h3> Please upload file 😘:</h3>
-            <form action = "scripts/php/get post file.php/extra/path" method="post" enctype="multipart/form-data">
+            <form action = "scripts/php/get post file.php" method="post" enctype="multipart/form-data">
                 <label>First Name: </label> <input type="text" name="first_name"><br><br>
                 <label>Surname: </label> <input type="text" name="last_name">
                 <br><br>
@@ -22,7 +22,7 @@
         <br>
         <div>
             <h3> Check uploaded files 🥰:</h3>
-            <form action="scripts/php/get post file.php/extra/path" method="get"enctype="multipart/form-data">
+            <form action="scripts/php/get post file.php" method="get"enctype="multipart/form-data">
                 <button type="submit" name="show_uploads" value="please">Show Uploaded files</button>
             </form>
         </div>
