+++ test/test.conf
@@ -3,6 +3,8 @@ server {
 	listen 8079;
 	index index.html;
 
+	error_page 404 /www/error_pages/400.html;
+
 	location / {
 		limit_except DELETE;
 	}
@@ -12,6 +14,11 @@ server {
 	}
 
 	location /index.html {
+		limit_except GET;
+		error_page 404 /www/error_pages/500.html;
+	}
+
+	location /ww {
 		limit_except GET;
 	}
